apps/api/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
asyncpg>=0.30.0
redis>=5.0.0
pydantic-settings>=2.0.0
httpx>=0.28.0
PyJWT>=2.10.0
cryptography>=44.0.0
